refactor: ToggleButton 네이밍 간략하게 수정
